created Ecorise users, app
