+++ Ecorise/Ecorise/settings.py
@@ -37,8 +37,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'trader',
-    'material_pricing',
+    'users',
+    
 ]
 
 MIDDLEWARE = [
