--- Ecorise/material_pricing/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-
-from .models import Material
-admin.site.register(Material)
