+++ Ecorise/users/admin.py
@@ -2,3 +2,6 @@ from django.contrib import admin
 
 from .models import Trader
 admin.site.register(Trader)
+from.models import Recycler
+admin.site.register(Recycler)
+
